Renaming according to this:
Parameter is variable in the declaration of function. Argument is the actual value of this variable that gets passed to function. ... When dealing with methods, the term parameter is used to identify the placeholders in the method signature, whereas the term arguments are the actual values that you pass in to the method